Switching to OnAllPluginsLoaded()
This makes it so you don't have to unload and load the plugin everytime the server restarts or map changes.

LuqS had mentioned there might be an issue with HookEvent(), which is why this helps, but not entirely sure.

Regardless, this fix helps me even though I don't fully understand it.

Co-Authored-By: Natanel Shitrit <65548905+Natanel-Shitrit@users.noreply.github.com>